refactor(feature_controller): Serialize Cilium features reconciliation
